+++ src/app/(beforeLogin)/i/flow/signup/page.tsx
@@ -1,4 +1,4 @@
-export default function Home(){
+export default function Page(){
 
     return(
         <main>
