+++ src/app/(beforeLogin)/layout.tsx
@@ -8,17 +8,17 @@ type Props = {
 export default function Layout({children,modal}:Props){
     return(
         <div className={styles.container}>
-            비포로그인 레이아웃
             {children}
             {modal}
         </div>
     )
 }
 
-// 주소가 localhost:3001일때는
-// children -> page.tsx
-// modal -> @modal/default.tsx
+// [ default.tsx ]
+// 1. 주소가 localhost:3001일때는
+// children -> page.tsx 
+// modal -> @modal/default.tsx 
 
-// 주소가 localhost:3001/i/flow/login 때는
-// children -> i/flow/login/page.tsx
-// modal -> @modal/i/flow/login/page.tsx
+// 2. 주소가 localhost:3001/i/flow/login 때는
+// children -> i/flow/login/page.tsx 
+// modal -> @modal/i/flow/login/page.tsx
